Move build.py to worker module
Building should only be done by workers
